make client instance a singleton to share across the app
